feat: branch out to have one file for each mapper and adapters
